fix: change prices double to int.

+++ src/DTOs/Product/ProductDtoForAdmin.cs
@@ -26,7 +26,7 @@ namespace api.src.DTOs.Product
         /// <summary>
         /// Atributo que representa el precio de un producto.
         /// </summary>
-        public required double Price { get; set; }
+        public required int Price { get; set; }
 
         /// <summary>
         /// Atributo que representa la cantidad de stock de un producto.
