+++ src/DTOs/Purchase/TicketDto.cs
@@ -17,7 +17,7 @@ namespace api.src.DTOs.Purchase
         /// Atributo de tipo double que representa el precio total de la compra.
         /// </summary>
         [Required]
-        public double Purchase_TotalPrice { get; set; } = 0;
+        public int Purchase_TotalPrice { get; set; } = 0;
 
         /// <summary>
         /// Atributo de tipo List<SaleItemDto> que contiene los artículos de venta asociados a este ticket.
